requirements.txt
streamlit
mysql-connector-python
pymongo
pandas
sqlalchemy
plotly
matplotlib
seaborn
dotenv